[531919] Move the tools computation out of the UI
- Add a new concept of ToolInstance in viewpoint model to contains tool
information including filter, visibiity and enablement status. Tools are
now available in the UIState of a DDiagram and not in the PaletteManager
anymore.

- A new component ToolManagment takes care of updating tools available
in the UIState. It is available from DiagramPlugin.

-PaletteManager now uses the UIState tool information of a DDiagram to
construct corresponding graphical palette. It registers itself as a
listener of ToolManagement to be aware of tool changes to launch an
update.

Bug: 531919
Change-Id: If26a1503ebd1018c9f10b4a3c968d82115a1adab
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>