[533988] Initialize text, line and fill colors in appearance tab
When clicking on text, line and fill color button in appearance view
when a node, edge, label or diagram is selected the colors displayed are
now the black to white colors scale, then the fixed user colors defined
in the VSM and then the other predefined Sirius colors following the
rainbow colors.
The number of columns is fixed to 10 and the maximum number of displayed
buttons is 50.

Bug: 533988
Cherry-picked-from: 533537
Change-Id: Id8a4da572a9ec7410a1d42291ad1acfd9125ffa1
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>