[533537] Initialize text, line and fill colors in appearance tab
When clicking on text, line and fill color button in appearance view
when a node, edge, label or diagram is selected the color displayed are
now the black to white color scale, then the fixed user color defined in
the vsm and then the other predefined Sirius color following the rainbow
color.
The number of column is fixed to 10 and the maximum number of displayed
buttons is 50.

Bug: 533537
Change-Id: I8218f7dc36c641baf181d07c54c5bc4da75f13b3
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>