[432400] Remove DRefreshable.refresh()
Remove the refresh() EOperation from the metamodel, and either remove
or move the manual code that existed in the various implementations.
This is part of the effort to reduce the amount on non-generated code
in the metamodel implementations, and the dependencies this can
bring (which prevents cleanly isolating the metamodels).
See bugs #531396 and #531282 for more context.

The DRefreshable type itself, which is empty after this change and
thus becomes just a "marker interface", will be removed later.

DRefreshable.refresh() was implemented by: DRepresentation, DView,
DRepresentationElement, and Style. Each case is treated separately.

For concrete DRepresentations (e.g. DDiagramSpec, DTableSpec, etc.),
all the implementations where exactly the same and delegated to
DialectManager.refresh(), so all call sites which target
DRepresentations now simply call that directly.

For DView, the only call that can end up in DViewSpec.refresh() is in
DeleteSemanticElementTest.testSemanticElementDeletionRemovesDiagram().
DViewSpec.refresh() does some actual work, so just to be safe for now,
the code has been kept but moved in a new internal
DViewHelper (outside of the metamodel's code) and called explicitly
from the test. It's possible this will be completely removed at a
later time.

For DRepresentationElement and Style, although these types are shared
by all dialects, the only concrete implementations and the only uses
of the operation are in diagrams, and even more precisely the only
mechanism through which the operation is called is
oes.diagram.tools.api.command.view.RefreshSiriusElement. All the
concrete implementations have been moved into an internal switch-based
helper inside RefreshSiriusElement.
RefreshSiriusElement.refresh(DRefreshable) is made public so that
client code which used to call DRefreshable.refresh() can be adapted.

After the move, this leaves a number of *Spec classes empty (they only
existed to override refresh()), so these are removed completely.

Bug: 432400
Change-Id: I209b5bf74bca8cc640c8b02c27ee5897a1a5a436
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>