[533002] Fix two NPE when Link Notes are used in fragmented models
The first NPE happened when using the link note tool on a diagram in an
aird fragment. The second NPE happened when double-clicking on a link
note tool that targets a diagram in an aird fragment.

Both exceptions were caused by the wrong notion that a session URI can
be retrieved by looking at the resource uri of a view's root container
object.

Bug: 533002
Change-Id: I6e07336de6f6a4ec45379130665610864db4769b
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>