[533854] Fix two NPE when Link Notes are used in fragmented models
The first NPE happened when using the link note tool on a diagram in an
aird fragment. The second NPE happened when double-clicking on a link
note tool that targets a diagram in an aird fragment.

Both exceptions were caused by the wrong notion that a session URI can
be retrieved by looking at the resource uri of a view's root container
object.

Bug: 533854
Change-Id: If5ee272c7877fd77926ffd4e653e051eda410e99
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>