[533996] Launch the validation in a valid context
The previous commits for bug 533996 and bug 509970 imply to change the
validation code. Indeed, the revalidate called during
ValidationFixResolution must be done on the editor only if this editor
is not the dummy editor opened by the GoTo marker. This problem appears
when a quick fix is applied on a session with no editor opened.
This commit also adds quick fix tests (to reveal the problem and to
avoid regressions).

Bug: 533996
Change-Id: Id33a5d6e284bf50d6fee2b15d4b936d23d3c2629
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>