[481386] Fix regression detected by a new test
The tests added in GoToMarkerTraceabilityWithUserInteractionTest for bug
534273 reveal a regression in the code of AbstractDDiagramConstraint
modified to fix bug 481386.
Indeed, if a semantic element is deleted, with the new code there is a
RuntimeException when calling
InterpreterRegistry.getInterpreter(EObject) on this semantic element.
This case is now handled in getFailingRules(EObject), before initial
commit for bug 481386, it was indirectly done in
checkRulesFromActivatedViewpoints(EObject, DDiagram) by calling
SessionManager.INSTANCE.getSession(EObject) and by checking if the
session was null. Now with this commit, the eObject that are not 
contained in a resource are ignored from the validation.
This commit also makes the tests from
GoToMarkerTraceabilityWithUserInteractionTest more reliable. I noticed
locally sometimes failure on testTraceabilityWithClosedSession.

Bug: 481386
Change-Id: I68c7990ce1332a88dbec2c39c7e656f6c4922c54
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>