[releng] Add org.apache.xalan in orbit.tpd
It is not really needed for Sirius TPs as it is automatically added
because of "planner" mode ("with requirements" flag in targetplatform
file). But if other TPs reuse this tpd file and is in "slicer" mode,
there is a dependency problem: "Missing requirement:
org.apache.batik.dom 1.9.1.v20180313-1559 requires 'package
org.apache.xml.utils 2.7.1' but it could not be found". This problem is
fixed by this commit.

Change-Id: Ic151884f13d4750218086534ab36ca4e39677aaf
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>