[releng] Use Jetty 9.4.10 for the build
We do not really depend on a particular version of Jetty, but pointing
to a specific update site instead of the global "Jetty 9.x" aggregate
repo (which contains a lot of entries) speeds up target resolution.

The specific version chosen, 9.4.10, corresponds to the one used by
the Eclipse Platform itself for Photon.

Change-Id: Iefe641b38711c076622b42890aa8ae7be0ba0b01
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>