[test] Fix some tests by changing the way to build the target platform
Since the addition of tests for bug 500253, they have always failed when
they are launched from Tycho. After analysis, we observed that the
target platform built by TestsUtil.setTargetPlatform() is wrong. Indeed
the Sirius plug-ins appear twice: one from the built version by Tycho
and one from the original git repository. The one from the git
repository is retained by the eclipse runtime launched for tests and it
provided wrong classes path with something like
"bin.org.eclipse.sirius..." instead of "org.eclipse.sirius...". The
consequence is that the compilation of project inside the  eclipse
runtime launched for tests is wrong and so some tests fail.
This commit excludes the plug-ins from the original git repository.

Change-Id: Ibf5df4bb1ef00513f11c7e673ba3ede2f7a0a2f3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>