[536009] ExportAction doesn't open Dialog Window
The errors are now handled in the callers
(AbstractExportRepresentationsAction and
ExportRepresentationsWorkspaceModifyOperation) through the ExportAction
execute action.

Since the execute action can only throw InvocationTargetException and
InterruptedException, the image generation errors are wrapped inside an
InvocationTargetException and callers have to properly handle the
case if there is some specific treatments to do.

Bug: 536009
Change-Id: Ib6e1ed953b4021cab28a153ca1cc656809d61b4e
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>