[537220] Fix deadlock and concurrent writing
Fix a deadlock occurring when saving a VSM with aird editor opened as
well as a representation.

This fix does not refresh anymore aird editor models block tree viewer for
each resource set change.

This old behavior introduced to see label changes of a model tree
element in the aird editor models block after a modification made
elsewhere seems to be obsolete.

Also fix concurrent writing access of 537219 happening for the same
reason.

Random thread order execution is the reason why we have the first or
second issue.

Bug: 537220
Bug: 537219
Change-Id: I3076858ff4d0688022f2a4c3dae281ad1dad0582
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>