[535861] Blank regeneration of VSM editor
This commit only launches a generation of VSM editor without any MM
change (except the date and the last "carriage return" of the
Copyright).
All chain files have been launched, then the action "Source/Clean up..."
has been launched twice and finally the action "File/Convert Line
Delimiters to/Unix" has been launched on all modified files.

List of changes caused by a previous missed generation:
* viewpoint:
    * DAnnotation.references: from bug 493353 - Add a mean to store
custom data on DRepresentationDescriptor --> The generation was propably
not done because the DAnnotation are not displayable in the editor.
* diagram:
    * EnumLayoutOption: Not missed but replaced by another class. It is
preferable to not delete unused classes to facilitate the next regen and
identify the delta.
    * EnumSetLayoutOption from bug 509070 - Update generic layout
algorithm integration to handle options.

Bug: 535861
Change-Id: I5b52216f3f8130886b530371f92a65d16ccc7d09
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>