[527109] Layer and filter visibility handling
In show/hide mode, when a double click is done on an invisible element
that is not visible because some filter(s) are activated, then a confirm
dialog is opened and asks user if he wants to continue by removing the
filter.

Also when the element is not visible because the layer showing it is not
activated, then a confirm
dialog is opened and asks user if he wants to continue by activating the
layer.

Bug: 527109
Change-Id: Idbbaedede002e90529b6bbdb07a1f42c6f8e842d
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>