[525261] Add technical id on Sirius meta-model
* Add specification
* Every classes of Sirius meta-model that does not inherit from a Sirius
class, now inherits from IdentifiedElement. Classes from Description
packages are not concerned.
* IdentifiedElement has a uid EAttribute of type String that represents
the unique id of the element. This attribute is valued in the
constructor. DRepresentation:uid attribute has been removed as it is
replaced by IDentifiedElement:uid
* AirdResourceImpl overrides attachedHelper method to uses uid of
IdentifiedElement as id in the id <=> EObject cache of the
AirdResourceImpl resources
* Add a migration participant to update IdentifiedElement.uid from the
previously serialized xmi:id and update
DRepresentationDescriptor:repPath
* Modify RepresentationsFileXMIHelper.getId() to return null for
IdentifiedElement instances and avoid to serialize an xmi:id in favor of
the uid which will also be used as href.
* Make Sirius EqualityHelper overrides Ecore.EqualityHelper so that
EAttribute of type ID=true are ignored to know if two EObjects are
equal.
* Add a test to ensure that all meta-classes, that should inherit from
IdentifiedElement, do it effectively.
* Add a test for migration

Bug: 525261
Change-Id: I2a9bc8d1abae7b63a4d100d0db71c7bfe8476aef
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>