[536743] Add support for the cost directive
Add a cost directive which will be used to prevent users from creating
queries which would request a massive amount of data. Each query will
have a maximum cost to respect (default: 100 points). Navigating any
multi-valued reference will have a cost defined thanks to the cost
directive.

If a query has a cost superior to the maximum cost, it will partially or
totally fail (depending on the shape of the query). Only the part of the
query which make it go above the maximum cost will fail. Additional
fields will have a cost in the future if they are creating performance
issues.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=536743
Change-Id: If9e72d1812936d8e6a9f266c874a4d9d0d249dee
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>