[537321] Improve the customization of the server
Change the behavior of the product to remove the autostart and remove the
autostart of the product itself from the Docker container. It will allow others
to create new Docker containers which can install additional features into the
product before starting it.

Previously they would have to create a new Docker container which would start
by stopping the autostarted server to install anything inside. This operation
was complex and useless.

Bug: http://bugs.eclipse.org/bugs/show_bug.cgi?id=537321
Change-Id: I25475ef257fad9b1bee44eaa0fe11a2ea3300e80
Signed-off-by: Hugo Marchadour <hugo.marchadour@obeo.fr>