[537232] Fix deadlock with workflow page
The Jetty Browser refresh calls SiriusServerBackendFilter in a separate
thread for each request coming from each resource set event.
This call constructs the Workflow page and during the construction,
Viewpoints proxys can be resolved at the same time leading to a deadlock
in EMF transaction level that is not thread safe.

This patch makes the browser to refresh only one time when the following
event have occurred:
- a representation change
- a semantic change
- a view selection change
- a VSM update
- a resource replacement

Bug: 537232
Change-Id: I76335690f8e71f4ec8fc03943112faf820748d7c
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>