[525466] Fix error marker persisting after modeling project conversion
In ModelingProjectManagerImpl.doAddModelingNature, the error marker
about the lack of aird is now created only if the exception is not the
one about zero representations file. The removal in this situation was done
by DefaultModelingProjectResourceListener.resourceChanged which was 
sometimes called before the marker creation.

A synchronized access to ModelingProject instance has been added in
ModelingProjectManagerImpl.doAddModelingNature to compute the main
representations file correctly and avoid modifications of the valid
attribute by other thread during the computation.

Bug: 525466
Change-Id: Ic1d2ffe1b79331bc0318459af0219237d0a36146
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>