[537513] Add support for CORS preflight requests for GraphQL
Add the ability for a piece of JavaScript loaded from a domain "A" to call the
GraphQL API if the server is located on a domain "B". In such situation, an
OPTIONS request is necessary first in order to determine if the server allows
such a call.

This commit adds support for both HEAD and OPTIONS request along with the
ability to respond without any body instead of a body with null inside.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=537513
Change-Id: Ia8a186c88a7e30f4ace62345729afc3e11949f02
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>