[531396] Clear dependencies to Sirius custom code.
Simplify some Viewpoint model code to reduce dependencies to Sirius
code. These Viewpoint model class are used in Metamodel code and will be
moved into Metamodel plug-in in an other commit.
Two Spec class are deleted because they are not used.
A new Interface IMigrationHandler (and extension) has been created (will
be moved in MetaModel plug-in) to use migration process into Metamodel
plug-in. 

Bug: 531396
Change-Id: I80dfb1adddf359ef20d786848dfa476017f13cf9
Signed-off-by: jmallet <jessy.mallet@obeo.fr>