[536743] Add support for mutation in the Sirius domain
The GraphQL API now provides mutation for the Sirius domain. As such, it
is now possible to activate and deactivate viewpoints and to create
representations (diagram only for now). We can also retrieve the list of
activated viewpoints on a projet.

Additional tests have been contributed along with some improvements to
the GraphQL schema implementation.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=536743
Change-Id: Ie1a3fc608667706d73c4e1f004719e924d871639
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>