[537781] Move the frontend code of the workflow into another plugin
Now that the code of the frontend can be compiled with a specific path
for the deployment of the code. The code of the workflow can be moved to
another plugin which will expose it over HTTP on the following path:

http://localhost:PORT/workflow

As a result, the code of the frontend of the Sirius web application will
be available on / while the frontend code of the workflow will be
accessible on /workflow. It will be easier to modify the code of the
frontend without any impact on the workflow and its required IE11
compatibility.

Bug: https://github.com/eclipse/sirius-components/issues/93
Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=537781
Change-Id: Id6d90635126c2e1710bdf7b488281b75784ebf16
Signed-off-by: Stéphane Bégaudeau <stephane.begaudeau@obeo.fr>