[538376] Change eAdapters() of EdgeMappingImportWrapper
EdgeMappingImportWrapper must return the eAdapters of the
wrapped EdgeMappingImport and not the eAdapters of the imported
EdgeMapping.
This is the same logic than for eResource, eContainer, ....
This is a part of the problem. With this commit, there is no longer the
"ping-pong" detected with team usage of Sirius. The main part of the
problem will be fixed in the following commit.

Bug: 538376
Change-Id: I6f90eedf44d066c721335a97fb2134b757b58d8e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>