[538376] Do not use EdgeMappingImportWrapper for DEdge.actualMapping
The EdgeMappingImportWrapper was created to allow to virtually use an
EdgeMappinpImport as mapping of a DEdge.
This is authorized as the getMapping() is an operation and the result is
never serialized. But the getActualMapping() should never returns an
EdgeMappingImportWrapper. The current behavior is "OK" locally as there
is a notification not sent in DEdgeSpec. But in team environment ODTE or
T4C, this is problematic as a notification, a CDOSetFeatureDeltaImpl to
be more precise, is created during the refresh, sent during commit and
processed by remote during invalidation. This causes problems as
EdgeMappingImportWrapper is not really an element model and furthermore
it is compared later with its EdgeMappingImport, that is not equal, and
throws another notification.

Bug: 538376
Change-Id: I5b30e8aa445e04287d6313f15a9a2a70f835a7aa
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>