[test] Add 3 swtbot sequence tests as unreliable
These tests fail since a long time (since the switch to a new version of
SWTBot and/or Oxygen). They fail only on some environment and are not
in failure because of a bug but more because of an environment
(incompatible GTK version with Oxygen and/or new version of SWTBot).
They have been added as unreliable to be visible only when we launch
unreliable tests.

Change-Id: I8f884f9efd0a2fa1e17f12c773444cf2d442ec34
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>