[525261] Make AirDResourceImpl::useIDAttributes return false
The cache used to get an element from its ID is idToEObjectMap.
This map is updated in attachedHelper and detachedHelper and should be
always up to date.
If this.getEObject(String uriFragment) is called and uriFragment
corresponds to an object that is not in the resource then, if
useIDAttributes=true, the whole resource content is iterated.
To avoid that, useIDAttributes return false.

Bug:525261
Change-Id: I4910c5075066a1f8998e9be4722058bd877c967f
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>