[test] Add several SwtBot tests as unreliable
These tests fail since a long time on some IC servers (since the switch
to a new version of SWTBot). They fail only on some environment and are
not in failure because of a bug but more because of an environment
(incompatible GTK version with Oxygen and/or new version
of SWTBot).
They have been added as unreliable to be visible only when we launch
unreliable tests. They will probably fixed by gerrit [1] in Sirius when
gerrit [2] in SwtBot will be merged.

[1] https://git.eclipse.org/r/#/c/127354/
[2] https://git.eclipse.org/r/#/c/127340/

Change-Id: Id1f6d783e6f3abe89814e13a9121b1eb8f8bf1c5
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>