[539550] Migrate broken Note, Text and NoteAttachment
A migration participant has been be added in Sirius to migrate .aird
from previous versions and fix invalid views if some of them are found
during load.

In some conditions, we can get aird files containing invalid Note, Text
or Note Attachment :
 - their element reference is unset (isSetElement() returns false,
value: null) and there is nothing serialized
 - whereas valid Note, Text and Note Attachment have an element
explicitly set to null isSetElement() returns true, the xml contains
<element xsi:nil="true"/>

This is also true for the BasicDecorationNode of Text/Note.

The element reference is explicitly st to null by GMF factories when the
user use the corresponding tools:
 - org.eclipse.gmf.runtime.diagram.ui.view.factories.BasicNodeViewFactory.createView()
in line 107
 - org.eclipse.gmf.runtime.diagram.ui.view.factories.ConnectionViewFactory.createView()
in line 104

If this null element is not serialized, the effect can be to have
invalid labels (the label of the current diagram for example) displayed
as label of the Note. This comes from the implementation of
org.eclipse.gmf.runtime.notation.impl.ViewImpl.getElement()

Bug: 539550
Change-Id: I28b410094e61815e0cc70c07f32966ba853aaff6
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>