[540056] Improve Link Note delete lifecycle
Link notes are no longer automatically deleted when the linked
representation is deleted. This prevents unnecessary locks in
collaborative sirius and the user won't lose the note text. If
a linked representation is deleted, the note will indicate that
it was linked to a representation that has been deleted.

Bug: 540056
Cherry-picked-from: 535648
Signed-off-by: Felix Dorner <felix.dorner@gmail.com>
Change-Id: Ib2583ee8ce4cdf9cab67db194bf64b1ac1446fd6