[540475] Force version of Sonar to be used for Sonar analysis
The version currently used by default, 3.5.0.1254, is no longer
compatible with the version of SonarQube server used in Eclipse infra.
So as said in sonarqube documentation [1], this commit forces the
version 3.3.0.603 to be used:
"From maven-sonar-plugin 3.4.0.905, SonarQube < 5.6 is no longer
supported.
If using SonarQube instance prior to 5.6, you should use
maven-sonar-plugin 3.3.0.603."

[1] https://docs.sonarqube.org/display/SCAN/Analyzing+with+SonarQube+Scanner+for+Maven

Bug: 540475
Change-Id: I73ca4756bf2dd57e61540ba86e60bdc78c48f18b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>