[541140] Do not refresh the GMF model during export as image action.
* The GMF refresh was introduced to avoid ClassCastException if the GMF
model is not consistent with the Sirius model. To avoid modifying the
model during the export as image, we no longer perform this refresh. If
a ClassCastException occurs because of a refresh issue, an exception is
raised and caught later to warn the end user that some of its diagrams
might need a manual refresh before exporting them.

Bug: 541140
Change-Id: Ie78a4a6df967a0477caab01ad6cc008bc907826a
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>