[cleanup] Remove useless exception handling code
The *only* type of checked exception that export() can throw is
SizeTooLargeException, which is a subtype of CoreException, so it
makes no sense to catch the more general CoreException and then test
for SizeTooLargeException.

Once the "instanceof" test is removed, the whole catch block becomes:

    } catch (SizeTooLargeException exception) {
        throw exception;
    }

which can obviously be removed completely.

Change-Id: I98956debc93205d53f4016cba67cfdc59aa055d0
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>