[541734] Catch possible exceptions if repDesc repPath is "invalid"
When the DRepresentationDescriptor.repPath URI is malformed or if it
does not correspond to any representation, it may raise an exception. We
consider that this case is a normal case and hence we catch the
exception so that the client get a null representation.

Bug: 541734
Cherry-picked-from : 541513 commit
cf22e4cd86025eeddbc393c2c88938147691ab7e
Change-Id: Ic2d7ec8719d173291ed7b1060d4086fd4a9cf84d
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>