[test] Attempt to fix SyncCall2Test.test_SyncCall_Insertion_Expand_Reparent
The test works fine locally, and other instances of the same code
pattern also work fine on the CI server, but this particular test
regularly fail on the CI server. This suggests the problem may be due
to a timing issue. Add an explicit waiting step to make sure the
selection is completely done.

Change-Id: I785611f055c5381f0579437bd85da0b217f5a815
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>