[543049] Fix SiriusCopier for non IdentifiedElement
Previously, the calls to EMF Copier methods were replaced by
SiriusCopier so that EAttribute that are Id=true, are not copied. It
was/is useful for IDENTIFIED_ELEMENT__UID EAttribute to ensure unicity
but is an issue for semantic object attributes. Indeed, in case of
copy/paste, the id attribute of the copied object may be used to
retrieve the original object.
Now, the SiriusCopier has been renamed SiriusCopierHelper and exposes
methods that hide internal implementation. The methods avoids copying
the attribute only for IDENTIFIED_ELEMENT__UID feature.

Bug: 543049
Change-Id: I4d1aec1b9512b3589f0e56fe741b0f9b8ce3fe4e
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>