[543252] Adds Hide and Show actions in visibility mode.
* The show and hide actions available by performing a double-click are
now present in the tabbar and contextual menu.
* The RevealOutlineElementsAction has been renamed in
RevealElementsAction.
* If an edit part is selected, the action delegate to this edit part to
retrieve the command to have the same behavior than the double-click
action.
* The show action is also displayed for hidden elements because of a
deactivated layer or filter.
* Complete the SWTBot test to check each action (tabbar and contextual
menu)
* Adds the show label action in the tabbar and make the show and hide
actions work together in the same way than the pin/unpin actions. For
instance, when clicking on show label, the hide label action is revealed
and the show label is disable if no other selected element has a hidden
label.
* Removes the old way to contribute the hide label action.

Bug: 543252
Change-Id: I873436740c9edb80e3275f77bd4a561dafd8aedd
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>