[test] Make test CustomizationPropertySectionsTests more reliable
Sometimes,
org.eclipse.sirius.tests.swtbot.editor.vsm.CustomizationPropertySectionsTests.testEReferenceCustomization()
fails because of a wrong number of elements ("The left list of available
elements should be of 5, i.e. all available style description elements
expected:<5> but was:<0>"). When it happened, I think that the popup is
not finished to initialize.
This commit adds a "wait until" instead of a simple assertion.

Change-Id: I1a5dd4a7dd1c8c85528c3f80229ead2304a76fbe
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>