[test] Make session opening during tests more reliable
Sometimes, the tests fail with the error "Timeout after: 1000 ms.: The
expected number of session was not reached.". In this case, the
screenshot shows that the session is not completely opened. The progress
status shows "Loading models: (5%)". This corresponds to
org.eclipse.sirius.ui.tools.internal.views.common.modelingproject.OpenRepresentationsFileJob.performOpenSession(IProgressMonitor)
just after the first "subMonitor.worked(1);". By adding a "join" on the
job family "AbstractRepresentationsFileJob.FAMILY" we ensure to wait the
end of the opening before continuing.

Change-Id: I76c794f697b5b00111e0c0db68206491e7a26285
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>