[547727] Fix viewpoint not referenced correctly by cross referencer
In some very rare circumstances, it can happen that viewpoints which
depend on each other (including at the OSGi level) are not loaded in
the correct order inside the registry's ResourceSet. For example
viewpoint A which references elements from B gets loaded before B.
This results in references from A to B to stay unresolved (proxy),
leading to hard to diagnose errors at runtime (e.g. missing mappings).

Forcing a full resolution every time a new viewpoint is loaded "fixes"
the issue, but has a significant performance cost at startup (when all
the viewpoints are discovered and loaded). This resolution is not done
by default as it only happens in rare circumstances in practice, but
can be enabled for users that need it by setting the
org.eclipse.sirius.viewpoint.registry.forceResolveOnLoad system
property to "true".

Bug: 547727
Change-Id: I72b0863f7002291df3c06fff1a4c5d92c0996baa
Cherry-picks: 544563
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>