[544563] Fix viewpoint not referenced correctly by cross referencer
In some very rare circumstances, it can happen that viewpoints which
depend on each other (including at the OSGi level) are not loaded in
the correct order inside the registry's ResourceSet. For example
viewpoint A which references elements from B gets loaded before B.
This results in references from A to B to stay unresolved (proxy),
leading to hard to diagnose errors at runtime (e.g. missing mappings).

Forcing a full resolution every time a new viewpoint is loaded "fixes"
the issue, but has a significant performance cost at startup (when all
the viewpoints are discovered and loaded). This resolution is not done
by default as it only happens in rare circumstances in practice, but
can be enabled for users that need it by setting the
org.eclipse.sirius.viewpoint.registry.forceResolveOnLoad system
property to "true".

Bug: 544563
Change-Id: Id123e80564070d961007fd29638aefcb25f0de79
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>