[545217] Remove bendpoints fails for edge with same source and target
RectilinearEdgeUtil computation now process edge with the same source
and target. The resulting edge has the source bendpoint moved to the
center of its side and the target is moved to the next side in a
clockwise order.
This fix on the Remove bendpoints action has also been applied to a
migration participant repairing bendpoints and to the ConnectionsFactory

Bug: 545217
Change-Id: Ic9bb2e29ca8e08545c56fd7c28d07e14b5759152
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>