[544834] Remove bendpoints fails for edge with same source and target
RectilinearEdgeUtil computation now process edge with the same source
and target. The resulting edge has the source bendpoint moved to the
center of its side and the target is moved to the next side in a
clockwise order.
This fix on the Remove bendpoints action has also been applied to a
migration participant repairing bendpoints and to the ConnectionsFactory

Bug: 544834
Change-Id: Ib8f1ed93be38c8f5bffd30faf2649f2b22944fc0
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>