[545594] Pb of edge visibility with collapsed compartment
By default, GMF does not consider edges on ContainerEditPart, parent of
the ShapeCompartmentEditPart, during the refreshConnections mechanism of
the ShapeCompartmentEditPart. But in Sirius, it is possible to have
such.
This commit adds the edges pointing to ContainerEditPart (as source or
as target, parent of the ShapeCompartmentEditPart, in the
refreshConnections mechanism.

Bug: 545594
Cherry-picked-from: 545533
Change-Id: I18054275645e4a38c7858c8c6857ce84c68c767b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>