[545853] Avoid a possible dead lock while selecting new created elements
In org.eclipse.sirius.ui.tools.internal.editor.SelectDRepresentationElementsListener.resourceSetChanged(ResourceSetChangeEvent),
we indirectly perform a Display.syncExec to retrieve the current active
editor (org.eclipse.sirius.common.ui.tools.api.util.EclipseUIUtil.getActiveEditor()).
As we are retaining the current EMF Transaction (the transaction lock is
released once the postCommit() is executed) that could potentially cause
a deadlock.
For instance, when performing a manual refresh on a diagram, new
graphical elements can be created. In the case of Capella for instance,
the org.polarsys.capella.core.ui.properties.CapellaDataListenerForPropertySections
is notified and schedule the refresh of the properties view within the
UI thread. This refresh try to acquire the lock on the transaction but
this one is kept by the Sirius refresh job and this job is waiting for
the ui thread in
electDRepresentationElementsListener.resourceSetChanged.
More generally, we should never wait for the UI thread in a thread that
has a lock on some resources.

Bug: 545853
Change-Id: I7c13e15eada6ab2f6ef3a446de5cc1546fd6eb45
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>