[546681] Fix NPEs when closing editors asynchronously
Patch https://git.eclipse.org/r/c/125716 on the platform cleans up
some fields which asynchronous callbacks used to access for closing
the editors. At the time the callbacks are executed, getPage() and
getWorkbenchWindow() now return null.

For DDiagramEditorImpl, the solution is the same that was used in GMF
Runtime 1.12.1 itself, but it is reproduced here to avoid the bug when
running with previous versions of GMF.

For tables & trees, we rely on our own helpers as the context if
different and we do not have access to all the methods needed for the
GMF solution.

Bug: 546681
Change-Id: I9a7831ea2643c7f7b8090633f524e2c035406fc8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>