[550366] Add action to export Sirius diagram into ELK graph
This commit adds an action to export Sirius diagram into ELK graph and
save it in the java temp folder using the diagram name as file name.
This action can be launch by using "Export diagram as ELK Graph"
contextual menu on a diagram.
This action is added in a new plug-in to have it only in specific
environment (development env for example).
If no ELK layout is associated with the diagram, a default
"org.eclipse.elk.layered" will be added during the export.

Bug: 550366
Cherry-picked-from: 548817
Change-Id: I49deb4849ec74b10466f6ab13e1ee71e756e6380
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>