[550389] Convert Sirius diagram to ELK graph - Wrong node location
There was wrong location of node or container contained in container.
This delta corresponds to the insets of container. It is visible with a
space between the node and the edge in Layout Graph view. There were 2
mistakes:
    * The insets was added to the node, the child, and not to the
container.
    * The location must be set by ignoring the insets. Indeed, the
coordinates of an ElkNode must not considered insets. It is used by the
algorithm only.

Bug: 550389
Cherry-picked-from: 548820
Change-Id: I67a3a9cb07c0d7497bcac2e3d02c707a3b47c0fe
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>