[546681] Allow Guava up to v27.1 (which will be in 2019-06)
I checked that Sirius is indeed compatible with Guava 27.1 by using
"[27.1,27.2)" as a range and building against a recent I-build of
Orbit: no compilation errors. I have not run the tests on the results,
but the Guava APIs are usually stable enough that if what we use still
exists it very probably still behaves the same.

We can not be too strict and force the use of 27.1 as we must still be
compatible with earlier versions of Eclipse, so we still allow for
21.0 which is the version in Oxygen, the oldest platform we still
build against.

Bug: 546681
Change-Id: I66f70f2e66b54941c18e66dd4ba8221ff7ea0443
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>