[546681] Use a more reliable way to close/hide the outline view
When the outline view is not already open/visible, the previous
technique causes a timeout and test failure. If a previous test in the
suite had already closed the view and failed to restore it (maybe
because itself failed for some other reason), this could cause long
sequences of test failures until another test explictly restored the
normal state.

Bug: 546681
Change-Id: I6deb516e2cf0a6e4c88c7d37d84ef9b60906b6ca
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>