[550368] Update metamodel and UI to allow usage of all ELK options
From UI perspective, options applying to nodes, labels and edges can now
be overridden from the VSM in addition to the options applying to
parent (the root of the diagram).

A new column Targets has been added in the dialog allowing to select
layout algorithm options to override.

The filter in the options dialog is now applied to columns Targets and
Type.

Bug: 550368
Cherry-picked-from: 547306
Change-Id: I1c7966add035af6d20321752bccbdb1eefe685a5
Signed-off-by: Pierre Guilet <pierre.guilet@obeo.fr>