[546681] Provide a way to identify a Session after dispose.
Compute the id of the session during its construction instead of on each
call to getID().
It corresponds to the URI of the sessionResource which is set in the
constructor and whose URI is never expected to change: the
sessionResource instance can be replaced after a reload, but this is
done after a check on its URI, so their is no need to recompute it
later.
This has been useful during test failures analysis after the move to
Eclipse 2019-06.

Bug: 546681
Change-Id: I788a0bb26da88e13059dfcf636507c4fab9b15f1
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>