[550366] Open access to methods of LayoutProvider
To allow retrieving the associated layout provider to a diagram and get
its associated layout configuration, some methods have been set to
public
instead of protected. It will be used for example, in the "elk.debug"
plug-in to make the transformation from a Sirius diagram to an ELK
graph.

Bug: 550366
Cherry-picked-from: 548817
Change-Id: I91106c2bd88f3dfd24b7e08ffc5d85fccb5b1db5
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>