[546681] Fix LinkNoteFragmentModelCreationTest and NoteCreationTest
The SWTBotGefEditor.directEditType(String) we use clearly documents
that it assumes direct edit mode is already enabled. In other contexts
simulating the key presses automatically triggers it, but apparently
when executed inside our Docker image, something in the graphical
stack (X11 server/window manager/version of Gtk) is different, and an
explicit "F2" is needed to trigger direct edit mode.

Bug: 546681
Change-Id: I5ca25d37a353f9c9de83e093555ec6d4921d2248
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>