[549054] Add a uri fragment cache to optimize the refresh.
During refresh, several checks are done on mappings during the analysis
of the existing elements and the candidates computation.

The mappings are compared using the haveSameURIFragment() method which
is recursive and use the uriFragment of each ancestor of the objects to
compare.

The cost of this computation was proportional to the number of elements
and the depth of the mapping hierarchy. The cache allows to compute the
uri fragment of each mapping only once.

Bug: 549054
Change-Id: If37ed43b4c1d1c3046deba2d3db30f33285d7929
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>