[549054] Change the behavior and documentation of isInActivatedLayer
When the given parent diagram is not null, we now directly used it. In
the previous implementation, this given parent diagram was only used in
the case of getParentDiagram() returning null.

In the previous implementation even if a known parent diagram was given
as parameter, the DDiagramElement::getParentDiagram() was called whereas
it might already have been used to compute the known parent (see
isInActivatedLayer(DiagramMappingsManager, DDiagramElement) or other
callers).

Bug: 549054
Change-Id: I66f8ee97f2fa7acadf23482d135abccf6be7ab9d
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>